Update `changeset-derive` command to include review relations with bounds parameter (#5455)
* Update ElementToGeometryConverter::convertToGeometry() to include review relations so that they can be bounded
* Include restriction relations in the bounds checking
* Sonar findings and formatting
* Copyright

+++ hoot-core/src/main/cpp/hoot/core/ops/MetadataOp.cpp
@@ -56,7 +56,7 @@ void MetadataOp::apply(std::shared_ptr<OsmMap>& pMap)
   if (_datasetIndicator.first.length() == 0)
     return;
 
-  _numProcessed = static_cast<size_t>(_pMap->getElementCount());
+  _numProcessed = _pMap->getElementCount();
 
   _apply();
 }
